+++ packages/client/package.json
@@ -22,6 +22,7 @@
     "@mui/icons-material": "^5.11.16",
     "@mui/lab": "^5.0.0-alpha.131",
     "@mui/material": "^5.13.1",
+    "@mui/x-data-grid": "^6.5.0",
     "formik": "^2.2.9",
     "graphql": "^16.6.0",
     "jwt-decode": "^3.1.2",
