+++ packages/client/src/graphql/auth/auth.graphql
@@ -1,9 +1,9 @@
 query users($projectId: ID!) {
-    users(projectId: $projectId) {
-        id
-        fullname
-        email
-        role
-        createdAt
-    }
+  users(projectId: $projectId) {
+    id
+    fullname
+    email
+    role
+    createdAt
+  }
 }
