+++ packages/client/src/layouts/layout.tsx
@@ -1,77 +1,35 @@
-import { DarkMode, GitHub, LightMode } from '@mui/icons-material';
-import { AppBar, Avatar, Box, Button, Drawer, IconButton, Toolbar, Typography } from '@mui/material';
+import { GitHub } from '@mui/icons-material';
+import { Avatar, Box, Container, IconButton, Toolbar, Typography } from '@mui/material';
 import React, { FC } from 'react';
-import { useSettings } from '@context/settings.context';
-import { Link } from 'react-router-dom';
 import { Paths } from '@constants/paths';
+import { SideBar } from '@components/side-bar/side-bar';
+import { SideListItem } from '@components/side-bar/side-list-item';
+import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
+import { faUsers, faUserPlus } from '@fortawesome/free-solid-svg-icons';
+import { Header } from '@components/header';
+import { Footer } from '@components/footer';
 
 export interface LayoutProps {
   children: React.ReactNode;
 }
 
 export const Layout: FC<LayoutProps> = ({ children }) => {
-  const { settings, setSettings } = useSettings();
-
-  const toggleDarkMode = () => {
-    setSettings({ ...settings, theme: settings.theme === 'light' ? 'dark' : 'light' });
-  };
-
-  const [open, setOpen] = React.useState(false);
-
-  const toggleDrawer = (open: boolean) => () => {
-    setOpen(open);
-  };
   return (
     <Box
       sx={{
-        minHeight: `100vh`
+        display: 'flex'
       }}
     >
-      <AppBar elevation={0} color="transparent" position="sticky">
-        <Toolbar sx={{ justifyContent: 'end' }}>
-          <IconButton edge="start" color="inherit" aria-label="menu" onClick={toggleDarkMode}>
-            {settings.theme === 'light' && <DarkMode color="info" />}
-            {settings.theme === 'dark' && <LightMode color="warning" />}
-          </IconButton>
-          <Button onClick={toggleDrawer(true)}>Open Drawer</Button>
-          <Drawer anchor={'top'} open={open} onClose={toggleDrawer(false)}>
-            <Box sx={{ width: 'auto' }} role="presentation" onClick={toggleDrawer(false)} onKeyDown={toggleDrawer(false)}>
-              <Box>
-                <Link to={Paths.USER_LIST}>Add Users</Link>
-              </Box>
-            </Box>
-          </Drawer>
-        </Toolbar>
-      </AppBar>
-      {children}
-      <Box
-        component="footer"
-        sx={{
-          display: 'flex',
-          zIndex: 10,
-          alignItems: 'center',
-          justifyContent: 'center',
-          position: 'fixed',
-          bottom: 0
-        }}
-      >
-        <Toolbar sx={{ display: 'flex', flexWrap: 'wrap', alignItems: 'center', justifyContent: 'space-between', width: '100vw' }}>
-          <Box
-            component="a"
-            sx={{ display: 'flex', flexDirection: 'row', alignItems: 'center', textDecoration: 'none', color: 'inherit', flexGrow: 1 }}
-            target="_blank"
-            href="https://sail.bu.edu"
-          >
-            <Avatar sx={{ width: 36, height: 36, mr: 1 }} src="https://sail.codes/img/s_logo.png" />
-            <Typography component="span" variant="h5">
-              SAIL
-            </Typography>
-          </Box>
-          <IconButton href="https://github.com/hicsail" target="_blank">
-            <GitHub />
-          </IconButton>
-        </Toolbar>
+      <Header />
+      <SideBar>
+        <SideListItem nodeId={Paths.USER_LIST} label="Users" path={Paths.USER_LIST} icon={<FontAwesomeIcon icon={faUsers} />} />
+        <SideListItem nodeId={Paths.USER_LIST} label="Invite" path={Paths.INVITE} icon={<FontAwesomeIcon icon={faUserPlus} />} />
+      </SideBar>
+      <Box sx={{ flexGrow: 1 }}>
+        <Toolbar />
+        <Container sx={{ mt: 4 }}>{children}</Container>
       </Box>
+      <Footer />
     </Box>
   );
 };
